Moved files to a separate package and fixed variable naming
