+++ src/main/java/school/faang/groupusers/User.java
@@ -1,4 +1,4 @@
-package school.faang;
+package school.faang.groupusers;
 
 import lombok.ToString;
 import lombok.Getter;
@@ -19,13 +19,13 @@ public class User {
     private final String address;
 
     public static Map<Integer, List<User>> groupUsers(List<User> userList) {
-        Map<Integer, List<User>> groupUsers = new HashMap<>();
+        Map<Integer, List<User>> groupedUsersByAge = new HashMap<>();
 
         for (User user : userList) {
             int userAge = user.getAge();
-            groupUsers.putIfAbsent(userAge, new ArrayList<>());
-            groupUsers.get(userAge).add(user);
+            groupedUsersByAge.putIfAbsent(userAge, new ArrayList<>());
+            groupedUsersByAge.get(userAge).add(user);
         }
-        return groupUsers;
+        return groupedUsersByAge;
     }
 }
